+++ src/main/java/net/frozenblock/wilderwild/worldgen/impl/feature/IcicleClusterFeature.java
@@ -19,6 +19,8 @@
 package net.frozenblock.wilderwild.worldgen.impl.feature;
 
 import com.mojang.serialization.Codec;
+import java.util.Optional;
+import java.util.OptionalInt;
 import net.frozenblock.wilderwild.worldgen.impl.feature.config.IcicleClusterConfig;
 import net.frozenblock.wilderwild.worldgen.impl.util.IcicleUtils;
 import net.minecraft.core.BlockPos;
@@ -34,8 +36,6 @@ import net.minecraft.world.level.levelgen.feature.DripstoneUtils;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.FeaturePlaceContext;
 import org.jetbrains.annotations.NotNull;
-import java.util.Optional;
-import java.util.OptionalInt;
 
 public class IcicleClusterFeature extends Feature<IcicleClusterConfig> {